Update README.md
Project to build a CLI Discussion Forum

Our version of the Discussion Forum will be non-graphical and entirely text-based. However, this project will guide you on how to build many of the functionalities that you may find on UpGrad’s class forum, such as: Register and log into a user account Create and upvote questions Reply to questions, upvote replies, and list all replies to a question Delete a question and its replies Goals of this assignment We’ve implemented some parts of the project for you, so you’ll need to implement the rest of the project on your own. More specifically, you will need to complete the following parts of the project: Question.java All of the constructors and methods in this class Reply.java All of the constructors and methods in this class The following methods in AbstractEntity.java getId setId autoGenerateId The following methods in UserActivity.java postNewQuestion seeAllQuestions deleteQuestion The following method in UpvoteServiceImpl.java getUpvote We have provided Detailed Instructions on how to write these classes and functions in the starter files that you can download from Downloads portion given below. You will find these detailed instructions on the next page.

By working on this project, you will get hands-on experience with creating larger software projects using the main pillars of OOP, namely, objects, classes, polymorphism, and inheritance. In addition, you’ll get a chance to read pieces of code written by professional software engineers and learn from them. All of which will help you to develop a strong fundamental knowledge of software engineering and become a skilled software engineer.

Note: Please make sure that you make changes in the starter file, only in the 5 files mentioned above.

Step-by-step instructions: Step 1: To Import the project, go to the ‘Open’ option, navigate to the folder that contains the project starter files, and open the folder. Step 2: To start working on the project, navigate to the units and util folders and find the classes and functions that you have to complete. Step 3: After you’ve completed the project, you can execute it to see the final result by following the instructions below: Click on the ‘Run...’ option under the ‘run’ menu